Add separate file for `DeriveChild` command.
