+++ dpe/src/commands/mod.rs
@@ -6,6 +6,7 @@ Abstract:
 --*/
 pub(crate) use self::initialize_context::InitCtxCmd;
 
+use self::derive_child::DeriveChildCmd;
 use crate::{
     dpe_instance::DpeInstance,
     response::{DpeErrorCode, Response},
@@ -14,6 +15,7 @@ use crate::{
 use core::mem::size_of;
 use crypto::Crypto;
 
+mod derive_child;
 mod initialize_context;
 
 #[derive(Debug, PartialEq, Eq)]
@@ -110,78 +112,6 @@ impl TryFrom<&[u8]> for CommandHdr {
             return Err(DpeErrorCode::InvalidCommand);
         }
         Ok(header)
-    }
-}
-
-#[repr(C)]
-#[derive(Debug, PartialEq, Eq)]
-#[cfg_attr(test, derive(zerocopy::AsBytes, zerocopy::FromBytes))]
-pub struct DeriveChildCmd {
-    pub handle: [u8; HANDLE_SIZE],
-    pub data: [u8; DPE_PROFILE.get_hash_size()],
-    pub flags: u32,
-    pub tcb_type: u32,
-    pub target_locality: u32,
-}
-
-impl DeriveChildCmd {
-    const INTERNAL_INPUT_INFO: u32 = 1 << 31;
-    const INTERNAL_INPUT_DICE: u32 = 1 << 30;
-    const RETAIN_PARENT: u32 = 1 << 29;
-    const TARGET_IS_DEFAULT: u32 = 1 << 28;
-
-    pub const fn is_internal_input_info(&self) -> bool {
-        self.flags & Self::INTERNAL_INPUT_INFO != 0
-    }
-
-    pub const fn is_internal_input_dice(&self) -> bool {
-        self.flags & Self::INTERNAL_INPUT_DICE != 0
-    }
-
-    pub const fn is_retain_parent(&self) -> bool {
-        self.flags & Self::RETAIN_PARENT != 0
-    }
-
-    pub const fn is_target_is_default(&self) -> bool {
-        self.flags & Self::TARGET_IS_DEFAULT != 0
-    }
-}
-
-impl TryFrom<&[u8]> for DeriveChildCmd {
-    type Error = DpeErrorCode;
-
-    fn try_from(raw: &[u8]) -> Result<Self, Self::Error> {
-        if raw.len() < size_of::<DeriveChildCmd>() {
-            return Err(DpeErrorCode::InvalidArgument);
-        }
-
-        let mut offset: usize = 0;
-
-        let mut handle = [0; HANDLE_SIZE];
-        handle.copy_from_slice(&raw[offset..offset + HANDLE_SIZE]);
-        offset += HANDLE_SIZE;
-
-        let mut data = [0; DPE_PROFILE.get_hash_size()];
-        data.copy_from_slice(&raw[offset..offset + DPE_PROFILE.get_hash_size()]);
-        offset += DPE_PROFILE.get_hash_size();
-
-        let flags = u32::from_le_bytes(raw[offset..offset + size_of::<u32>()].try_into().unwrap());
-        offset += size_of::<u32>();
-
-        let tcb_type =
-            u32::from_le_bytes(raw[offset..offset + size_of::<u32>()].try_into().unwrap());
-        offset += size_of::<u32>();
-
-        let target_locality =
-            u32::from_le_bytes(raw[offset..offset + size_of::<u32>()].try_into().unwrap());
-
-        Ok(DeriveChildCmd {
-            handle,
-            data,
-            flags,
-            tcb_type,
-            target_locality,
-        })
     }
 }
 
@@ -350,19 +280,19 @@ impl TryFrom<&[u8]> for TagTciCmd {
 }
 
 #[cfg(test)]
-mod tests {
+pub mod tests {
     use super::*;
     use crate::dpe_instance::tests::{SIMULATION_HANDLE, TEST_HANDLE};
     use crate::{DpeProfile, DPE_PROFILE};
     use zerocopy::{AsBytes, FromBytes};
 
     #[cfg(feature = "dpe_profile_p256_sha256")]
-    const TEST_DIGEST: [u8; DPE_PROFILE.get_hash_size()] = [
+    pub const TEST_DIGEST: [u8; DPE_PROFILE.get_hash_size()] = [
         1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25,
         26, 27, 28, 29, 30, 31, 32,
     ];
     #[cfg(feature = "dpe_profile_p384_sha384")]
-    const TEST_DIGEST: [u8; DPE_PROFILE.get_hash_size()] = [
+    pub const TEST_DIGEST: [u8; DPE_PROFILE.get_hash_size()] = [
         1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25,
         26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48,
     ];
@@ -371,13 +301,6 @@ mod tests {
         magic: CommandHdr::DPE_COMMAND_MAGIC,
         cmd_id: Command::GET_PROFILE,
         profile: DPE_PROFILE as u32,
-    };
-    const TEST_DERIVE_CHILD_CMD: DeriveChildCmd = DeriveChildCmd {
-        handle: SIMULATION_HANDLE,
-        data: TEST_DIGEST,
-        flags: 0x1234_5678,
-        tcb_type: 0x9876_5432,
-        target_locality: 0x10CA_1171,
     };
     const TEST_ROTATE_CTX_CMD: RotateCtxCmd = RotateCtxCmd {
         flags: 0x1234_5678,
@@ -408,15 +331,6 @@ mod tests {
         assert_eq!(
             CommandHdr::read_from_prefix(command_bytes).unwrap(),
             CommandHdr::try_from(command_bytes).unwrap(),
-        );
-    }
-
-    #[test]
-    fn try_from_derive_child() {
-        let command_bytes = TEST_DERIVE_CHILD_CMD.as_bytes();
-        assert_eq!(
-            DeriveChildCmd::read_from_prefix(command_bytes).unwrap(),
-            DeriveChildCmd::try_from(command_bytes).unwrap(),
         );
     }
 
@@ -630,23 +544,6 @@ mod tests {
         assert_eq!(
             GOOD_HEADER,
             CommandHdr::try_from(GOOD_HEADER.as_bytes()).unwrap()
-        );
-    }
-
-    #[test]
-    fn test_slice_to_derive_child() {
-        let invalid_argument: Result<DeriveChildCmd, DpeErrorCode> =
-            Err(DpeErrorCode::InvalidArgument);
-
-        // Test if too small.
-        assert_eq!(
-            invalid_argument,
-            DeriveChildCmd::try_from([0u8; size_of::<DeriveChildCmd>() - 1].as_slice())
-        );
-
-        assert_eq!(
-            TEST_DERIVE_CHILD_CMD,
-            DeriveChildCmd::try_from(TEST_DERIVE_CHILD_CMD.as_bytes()).unwrap()
         );
     }
